Expire granafa session when cookies expire.
At the moment, grafana session expires after 7 days by default. On the
other hand, keystone token usually expires earlier. This leads to all
the panels reporting "Failed to obtain Keystone token" warning message
while the user remains logged in to grafana dashboard. Ideally, the
token would be renewed automatically but since the cookie containing the
password expires, this is not possible. Hence, logging out the user is
the best options which happens automatically after expiring the grafana
session which is in essence what this commit does.

Closes: #6